jm - support h/w diminished root note
Previously it was calculating starting note based on a 7-note scale
We start h/w diminished on the 6th degree of w/h so that our h/w
scale shows as having a b9 and #9 and the w/h shows as having
dim7 and major 7. Maybe w/h should actually start on the major 7,
but then it would show as having a perfect 1 on a bb1? TBD